+++ src/attributecode/gen.py
@@ -36,6 +36,7 @@ from attributecode.util import to_posix
 from attributecode.util import UNC_PREFIX_POSIX
 from attributecode.util import unique
 from attributecode.util import load_scancode_json, load_csv, load_json, load_excel
+from attributecode.util import strip_inventory_value
 
 
 def check_duplicated_columns(location):
@@ -131,6 +132,7 @@ def load_inventory(location, from_attrib=False, base_dir=None, scancode=False, r
     """
     errors = []
     abouts = []
+    is_spreadsheet = False
 
     if base_dir:
         base_dir = util.to_posix(base_dir)
@@ -143,8 +145,10 @@ def load_inventory(location, from_attrib=False, base_dir=None, scancode=False, r
                 errors.extend(dup_cols_err)
                 return errors, abouts
             inventory = load_csv(location)
+            is_spreadsheet = True
         elif location.endswith('.xlsx'):
             dup_cols_err, inventory = load_excel(location)
+            is_spreadsheet = True
             if dup_cols_err:
                 errors.extend(dup_cols_err)
                 return errors, abouts
@@ -154,7 +158,14 @@ def load_inventory(location, from_attrib=False, base_dir=None, scancode=False, r
     try:
         arp_list = []
         errors = []
-        for component in inventory:
+
+        if is_spreadsheet:
+            # Only the .csv and .xlsx may have newline issue
+            stripped_inv = strip_inventory_value(inventory)
+        else:
+            stripped_inv = inventory
+
+        for component in stripped_inv:
             if not from_attrib:
                 arp = component['about_resource']
                 dup_err = check_duplicated_about_resource(arp, arp_list)
@@ -174,13 +185,15 @@ def load_inventory(location, from_attrib=False, base_dir=None, scancode=False, r
         if errors:
             return errors, abouts
     except Exception as e:
+        print("!!!!!!!!!!!!!!!!")
+        print(str(e))
         # TODO: why catch ALL Exception
         msg = "The essential field 'about_resource' is not found in the <input>"
         errors.append(Error(CRITICAL, msg))
         return errors, abouts
 
     custom_fields_list = []
-    for fields in inventory:
+    for fields in stripped_inv:
         # check does the input contains the required fields
         required_fields = model.About.required_fields